requirements.txt
# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.2  # For JavaScript-heavy sites

# Data processing
pandas==2.1.3
numpy==1.26.2

# API clients
httpx==0.25.2

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Data validation
pydantic==2.5.2

# Utilities
python-dotenv==1.0.0
click==8.1.7  # For CLI scripts

# Logging
loguru==0.7.2

# Data export
openpyxl==3.1.2  # For Excel export
xlsxwriter==3.1.9

# Rate limiting
ratelimit==2.2.1

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Visualization (optional)
matplotlib==3.8.2
seaborn==0.13.0

# Development tools
pytest==7.4.3
black==23.12.0
flake8==6.1.0
